Spliting elevator position change and zeroing elevator into 2 classes
I understand that this is unconventional, but my idea is to split up the objectives (changing elevator position and zeroing elevator) between two classes to simplify their usage to try and follow Single Responsibility Principle. @PaarkG, I also changed Motors to leadMotor and followerMotor, as well as incorporation of a Sequential Command group inside of my Command factory in ZeroElevator. Assuming that this fix is stupid, then I can simply re-implement the code into ElevatorSubsystem and delete the ZeroElevator child class.